Clean up code; formatting

+++ src/main/java/me/zeroeightsix/kami/util/ColourUtils.java
@@ -235,8 +235,8 @@ public class ColourUtils {
         }
 
         public int computeMSE(int pixR, int pixG, int pixB) {
-            return (int) (((pixR - r) * (pixR - r) + (pixG - g) * (pixG - g) + (pixB - b)
-                    * (pixB - b)) / 3);
+            return ((pixR - r) * (pixR - r) + (pixG - g) * (pixG - g) + (pixB - b)
+                    * (pixB - b)) / 3;
         }
 
         public int getR() {
